Making a ThemeHandler component. Not sure I'm on the right track.
