+++ public/index.html
@@ -9,7 +9,6 @@
     <link href="https://fonts.googleapis.com/css?family=Mina|Open+Sans" rel="stylesheet">
     <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
     <link rel="shortcut icon" href="%PUBLIC_URL%/favicon.ico">
-    <link rel="stylesheet" href="src/css/style.css">
 </head>
 
 <body>
